+++ src/Command/GenerateModelCommand.php
@@ -2,12 +2,11 @@
 
 namespace Krlove\EloquentModelGenerator\Command;
 
-use Illuminate\Config\Repository as AppConfig;
 use Illuminate\Console\Command;
-use Krlove\EloquentModelGenerator\Config;
+use Illuminate\Database\DatabaseManager;
 use Krlove\EloquentModelGenerator\Generator;
+use Krlove\EloquentModelGenerator\Helper\Prefix;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputOption;
 
 /**
  * Custom Type Definitions
@@ -23,27 +22,11 @@ use Jsor\Doctrine\PostGIS\Types\GeographyType;
  */
 class GenerateModelCommand extends Command
 {
-    /**
-     * @var string
-     */
+    use GenerateCommandTrait;
+
     protected $name = 'krlove:generate:model';
 
-    /**
-     * @var Generator
-     */
-    protected $generator;
-
-    /**
-     * @var AppConfig
-     */
-    protected $appConfig;
-
-    /**
-     * GenerateModelCommand constructor.
-     * @param Generator $generator
-     * @param AppConfig $appConfig
-     */
-    public function __construct(Generator $generator, AppConfig $appConfig)
+    public function __construct(private Generator $generator, private DatabaseManager $databaseManager)
     {
         parent::__construct();
 
@@ -56,52 +39,18 @@ class GenerateModelCommand extends Command
         Type::addType('geography', 'Jsor\Doctrine\PostGIS\Types\GeographyType');
     }
 
-    /**
-     * Executes the command
-     */
-    public function fire()
+    public function handle()
     {
         $config = $this->createConfig();
+        $config->setClassName($this->argument('class-name'));
+        Prefix::setPrefix($this->databaseManager->connection($config->getConnection())->getTablePrefix());
 
         $model = $this->generator->generateModel($config);
+        $this->saveModel($model);
 
         $this->output->writeln(sprintf('Model %s generated', $model->getName()->getName()));
     }
 
-    /**
-     * Add support for Laravel 5.5
-     */
-    public function handle()
-    {
-        $this->fire();
-    }
-
-    /**
-     * @return Config
-     */
-    protected function createConfig()
-    {
-        $config = [];
-
-        foreach ($this->getArguments() as $argument) {
-            $config[$argument[0]] = $this->argument($argument[0]);
-        }
-        foreach ($this->getOptions() as $option) {
-            $value = $this->option($option[0]);
-            if ($option[2] == InputOption::VALUE_NONE && $value === false) {
-                $value = null;
-            }
-            $config[$option[0]] = $value;
-        }
-
-        $config['db_types'] = $this->appConfig->get('eloquent_model_generator.db_types');
-
-        return new Config($config, $this->appConfig->get('eloquent_model_generator.model_defaults'));
-    }
-
-    /**
-     * @return array
-     */
     protected function getArguments()
     {
         return [
@@ -109,20 +58,8 @@ class GenerateModelCommand extends Command
         ];
     }
 
-    /**
-     * @return array
-     */
     protected function getOptions()
     {
-        return [
-            ['table-name', 'tn', InputOption::VALUE_OPTIONAL, 'Name of the table to use', null],
-            ['output-path', 'op', InputOption::VALUE_OPTIONAL, 'Directory to store generated model', null],
-            ['namespace', 'ns', InputOption::VALUE_OPTIONAL, 'Namespace of the model', null],
-            ['base-class-name', 'bc', InputOption::VALUE_OPTIONAL, 'Model parent class', null],
-            ['no-timestamps', 'ts', InputOption::VALUE_NONE, 'Set timestamps property to false', null],
-            ['date-format', 'df', InputOption::VALUE_OPTIONAL, 'dateFormat property', null],
-            ['connection', 'cn', InputOption::VALUE_OPTIONAL, 'Connection property', null],
-            ['backup', 'b', InputOption::VALUE_NONE, 'Backup existing model', null]
-        ];
+        return $this->getCommonOptions();
     }
 }