+++ src/Generator.php
@@ -2,98 +2,51 @@
 
 namespace Krlove\EloquentModelGenerator;
 
-use Krlove\EloquentModelGenerator\Exception\GeneratorException;
-use Krlove\CodeGenerator\Model\ClassModel;
+use IteratorAggregate;
+use Krlove\EloquentModelGenerator\Config\Config;
+use Krlove\EloquentModelGenerator\Model\EloquentModel;
+use Krlove\EloquentModelGenerator\Processor\ProcessorInterface;
 
-/**
- * Class Generator
- * @package Krlove\Generator
- */
 class Generator
 {
     /**
-     * @var EloquentModelBuilder
+     * @var ProcessorInterface[]
      */
-    protected $builder;
+    protected array $processors;
 
     /**
-     * @var TypeRegistry
+     * @param ProcessorInterface[]|IteratorAggregate $processors
      */
-    protected $typeRegistry;
-
-    /**
-     * Generator constructor.
-     * @param EloquentModelBuilder $builder
-     * @param TypeRegistry $typeRegistry
-     */
-    public function __construct(EloquentModelBuilder $builder, TypeRegistry $typeRegistry)
+    public function __construct(iterable $processors)
     {
-        $this->builder = $builder;
-        $this->typeRegistry = $typeRegistry;
+        if ($processors instanceof IteratorAggregate) {
+            $this->processors = iterator_to_array($processors);
+        } else {
+            $this->processors = $processors;
+        }
     }
 
-    /**
-     * @param Config $config
-     * @return ClassModel
-     * @throws GeneratorException
-     */
-    public function generateModel(Config $config)
+    public function generateModel(Config $config): EloquentModel
     {
-        $this->registerUserTypes($config);
+        $model = new EloquentModel();
 
-        $model   = $this->builder->createModel($config);
-        $content = $model->render();
+        $this->sortProcessorsByPriority();
 
-        $outputPath = $this->resolveOutputPath($config);
-        if ($config->get('backup') && file_exists($outputPath)) {
-            rename($outputPath, $outputPath . '~');
+        foreach ($this->processors as $processor) {
+            $processor->process($model, $config);
         }
-        file_put_contents($outputPath, $content);
 
         return $model;
     }
 
-    /**
-     * @param Config $config
-     * @return string
-     * @throws GeneratorException
-     */
-    protected function resolveOutputPath(Config $config)
+    protected function sortProcessorsByPriority(): void
     {
-        $path = $config->get('output_path');
-        if ($path === null || stripos($path, '/') !== 0) {
-            if (function_exists('app_path')) {
-                $path = app_path($path);
-            } else {
-                $path = app('path') . ($path ? DIRECTORY_SEPARATOR . $path : $path);
+        usort($this->processors, function (ProcessorInterface $one, ProcessorInterface $two) {
+            if ($one->getPriority() == $two->getPriority()) {
+                return 0;
             }
-        }
 
-        if (!is_dir($path)) {
-            if (!mkdir($path, 0777, true)) {
-                throw new GeneratorException(sprintf('Could not create directory %s', $path));
-            }
-        }
-
-        if (!is_writeable($path)) {
-            throw new GeneratorException(sprintf('%s is not writeable', $path));
-        }
-
-        return $path . '/' . $config->get('class_name') . '.php';
-    }
-
-    /**
-     * @param Config $config
-     */
-    protected function registerUserTypes(Config $config)
-    {
-        $userTypes = $config->get('db_types');
-        if ($userTypes && is_array($userTypes)) {
-            $connection = $config->get('connection');
-
-            foreach ($userTypes as $type => $value) {
-                $this->typeRegistry->registerType($type, $value, $connection);
-            }
-        }
+            return $one->getPriority() < $two->getPriority() ? 1 : -1;
+        });
     }
 }