+++ src/Model/BelongsToMany.php
@@ -2,55 +2,18 @@
 
 namespace Krlove\EloquentModelGenerator\Model;
 
-/**
- * Class BelongsToMany
- * @package Krlove\EloquentModelGenerator\Model
- */
 class BelongsToMany extends Relation
 {
-    /**
-     * @var string
-     */
-    protected $joinTable;
+    protected string $joinTable;
 
-    /**
-     * BelongsToMany constructor.
-     * @param string $tableName
-     * @param string $joinTable
-     * @param string $foreignColumnName
-     * @param string $localColumnName
-     */
-    public function __construct($tableName, $joinTable, $foreignColumnName, $localColumnName)
+    public function __construct(string $tableName, string $joinTable, string $foreignColumnName, string $localColumnName)
     {
         $this->joinTable = $joinTable;
         parent::__construct($tableName, $foreignColumnName, $localColumnName);
     }
 
-    /**
-     * @return string
-     */
-    public function getDefaultJoinTableName()
-    {
-        //return
-    }
-
-    /**
-     * @return string
-     */
-    public function getJoinTable()
+    public function getJoinTable(): string
     {
         return $this->joinTable;
-    }
-
-    /**
-     * @param string $joinTable
-     *
-     * @return $this
-     */
-    public function setJoinTable($joinTable)
-    {
-        $this->joinTable = $joinTable;
-
-        return $this;
     }
 }