+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 CHANGELOG
 =========
+2022-11-20, v2.0.1
+-----------------
+* Restore service autodiscovery (issue [#92](https://github.com/krlove/eloquent-model-generator/issues/92))
+* Actually register custom sql type mappings (issue [#97](https://github.com/krlove/eloquent-model-generator/issues/97))
+
+2022-02-19, v2.0.0
+-----------------
+* Minimum supported versions: PHP 8.0 and Laravel 9
+* Add command to generate models for all tables in the database
+* `backup` option replaced with `no-backup`
+* Add unit and integration tests
+* Simplified and reorganised codebase
+* Include `enum` type cast by default
+
+2022-02-10, v1.3.8
+-----------------
+* Add Laravel 9 support
+
 2021-03-07, v1.3.7
 -----------------
 * Add Doctrine DBAL 3.0 support